# airframe-codec: 3-5x Improvement of the codec lookup performance of MessageCodecFactory 