MessageCodecFactory used Surface case class as a cache key, which had a significant overhead. Improved the performance by using Surface.fullName as a cache key:

Before
```
[info] Benchmark                                     Mode  Cnt          Score          Error  Units
[info] RPCRequestBenchmark.rpcBodyOnly              thrpt   10  143043471.699 ± 72497486.069  ops/s
[info] RPCRequestBenchmark.rpcRequestNoCodec        thrpt   10      89063.233 ±      624.039  ops/s
[info] RPCRequestBenchmark.rpcRequestWithJsonSer    thrpt   10      74278.516 ±     1438.154  ops/s
[info] RPCRequestBenchmark.rpcRequestWithRPCMethod  thrpt   10      58004.750 ±      984.233  ops/s
```

After
```
[info] Benchmark                                     Mode  Cnt          Score        Error  Units
[info] RPCRequestBenchmark.rpcBodyOnly              thrpt   10  186913889.715 ± 966841.794  ops/s
[info] RPCRequestBenchmark.rpcRequestNoCodec        thrpt   10     346416.119 ±   4292.134  ops/s
[info] RPCRequestBenchmark.rpcRequestWithJsonSer    thrpt   10     204541.213 ±   3527.284  ops/s
[info] RPCRequestBenchmark.rpcRequestWithRPCMethod  thrpt   10     299616.662 ±   1119.058  ops/s
```